+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 Unreleased
 ----------
 
-### Added
+### Changed
+* Renaming project name from `PhingDiagram` to `PhingVisualiser`.
 
 
 v0.0.0